♻️(tests) move test infrastructure to proper locations
