+++ README.md
@@ -23,7 +23,7 @@ $ docker compose version
 ```
 
 > ⚠️ You may need to run the following commands with `sudo` but this can be
-> avoided by assigning your user to the `docker` group. See docker 
+> avoided by assigning your user to the `docker` group. See docker
 > [Documentation](https://docs.docker.com/engine/install/linux-postinstall/)
 
 #### Optional: pre-commit
@@ -62,7 +62,7 @@ Your Docker services should now be up and running 🎉
 
 You can create a basic demo site by running:
 
-    $ make demo
+    $ make seed
 
 Finally, you can check all available Make rules using:
