--- src/backend/demo/defaults.py
@@ -1,21 +1,0 @@
-"""Parameters that define how the demo site will be built."""
-
-NB_OBJECTS = {"documents": 1000, "services": 5}
-
-DEV_SERVICES = (
-    {
-        "name": "docs",
-        "client_id": "impress",
-        "token": "find-api-key-for-docs-with-exactly-50-chars-length",
-    },
-    {
-        "name": "drive",
-        "client_id": "drive",
-        "token": "find-api-key-for-driv-with-exactly-50-chars-length",
-    },
-    {
-        "name": "conversations",
-        "client_id": "conversations",
-        "token": "find-api-key-for-conv-with-exactly-50-chars-length",
-    },
-)
